requirements.txt
Flask
gunicorn
python-dotenv
sentence-transformers
pinecone-client
google-generativeai
langchain-docling
